fast dico for release - 1
