+++ docs/net8.0/DevFast.Net.Collection.Abstractions.Concurrent.Hashed.md
@@ -1,9 +1,9 @@
 #### [DevFast.Net.Collection](index.md 'index')
 
-## DevFast.Net.Collection.Abstractions.Concurrent.LookUps Namespace
+## DevFast.Net.Collection.Abstractions.Concurrent.Hashed Namespace
 
 [namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains necessary [interface](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface')s to implement various hash-based data-structures.
 
 | Interfaces | |
 | :--- | :--- |
-| [IFastDictionary&lt;TKey,TValue&gt;](DevFast.Net.Collection.Abstractions.Concurrent.LookUps.IFastDictionary_TKey,TValue_.md 'DevFast.Net.Collection.Abstractions.Concurrent.LookUps.IFastDictionary<TKey,TValue>') | Interface for a thread-safe collection of key-value pairs that can be accessed by multiple threads concurrently. |
+| [IFastDictionary&lt;TKey,TValue&gt;](DevFast.Net.Collection.Abstractions.Concurrent.Hashed.IFastDictionary_TKey,TValue_.md 'DevFast.Net.Collection.Abstractions.Concurrent.Hashed.IFastDictionary<TKey,TValue>') | Interface for a thread-safe collection of key-value pairs that can be accessed by multiple threads concurrently. |
