+++ docs/netstandard2.0/index.md
@@ -7,9 +7,9 @@ This assembly contains various well-tested performance oriented collections.
 | Namespaces | |
 | :--- | :--- |
 | [DevFast.Net.Collection.Abstractions](DevFast.Net.Collection.Abstractions.md 'DevFast.Net.Collection.Abstractions') | [namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains necessary [interface](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface')s to implement various data-structures. |
-| [DevFast.Net.Collection.Abstractions.Concurrent.LookUps](DevFast.Net.Collection.Abstractions.Concurrent.LookUps.md 'DevFast.Net.Collection.Abstractions.Concurrent.LookUps') | [namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains necessary [interface](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface')s to implement various hash-based data-structures. |
+| [DevFast.Net.Collection.Abstractions.Concurrent.Hashed](DevFast.Net.Collection.Abstractions.Concurrent.Hashed.md 'DevFast.Net.Collection.Abstractions.Concurrent.Hashed') | [namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains necessary [interface](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface')s to implement various hash-based data-structures. |
 | [DevFast.Net.Collection.Abstractions.Heaps](DevFast.Net.Collection.Abstractions.Heaps.md 'DevFast.Net.Collection.Abstractions.Heaps') | [namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains necessary [interface](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/interface')s to implement various Heap data-structures. |
-| [DevFast.Net.Collection.Implementations.Concurrent.LookUps](DevFast.Net.Collection.Implementations.Concurrent.LookUps.md 'DevFast.Net.Collection.Implementations.Concurrent.LookUps') | [namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains various hash-based data-structures. |
+| [DevFast.Net.Collection.Implementations.Concurrent.LookUps](DevFast.Net.Collection.Implementations.Concurrent.LookUps.md 'DevFast.Net.Collection.Implementations.Concurrent.LookUps') | |
 | [DevFast.Net.Collection.Implementations.Heaps](DevFast.Net.Collection.Implementations.Heaps.md 'DevFast.Net.Collection.Implementations.Heaps') | [namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains various Heap data-structures. |
 | [DevFast.Net.Collection.Implementations.Heaps.AbstractBase](DevFast.Net.Collection.Implementations.Heaps.AbstractBase.md 'DevFast.Net.Collection.Implementations.Heaps.AbstractBase') | [namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains [abstract](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/abstract 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/abstract') heap implementations. |
 | [DevFast.Net.Collection.Implementations.ReSizing](DevFast.Net.Collection.Implementations.ReSizing.md 'DevFast.Net.Collection.Implementations.ReSizing') | [namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains various Re-sizing strategies for data-structures. |
