+++ docs/netstandard2.1/DevFast.Net.Collection.Implementations.Concurrent.LookUps.md
@@ -2,8 +2,6 @@
 
 ## DevFast.Net.Collection.Implementations.Concurrent.LookUps Namespace
 
-[namespace](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace 'https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/namespace') contains various hash-based data-structures.
-
 | Classes | |
 | :--- | :--- |
 | [FastDictionary&lt;TKey,TValue&gt;](DevFast.Net.Collection.Implementations.Concurrent.LookUps.FastDictionary_TKey,TValue_.md 'DevFast.Net.Collection.Implementations.Concurrent.LookUps.FastDictionary<TKey,TValue>') | Represents a thread-safe collection of key-value pairs that can be accessed by multiple threads concurrently. |
