+++ src/DevFast.Net.Collection.Tests/Implementations/Concurrent/Hashed/FastDictionary.Enumerator.Test.cs
@@ -1,15 +1,15 @@
 ﻿using System.Collections;
 using DevFast.Net.Collection.Implementations.Concurrent.LookUps;
 
-namespace DevFast.Net.Collection.Tests.Implementations.Concurrent
+namespace DevFast.Net.Collection.Tests.Implementations.Concurrent.Hashed
 {
     [TestFixture]
-    public class FastDictionaryKeyEnumeratorTest
+    public class FastDictionaryEnumeratorTest
     {
         [Test]
-        public void FastDicionary_When_Empty_Returns_KeyEnumerable_Of_Zero_Count()
+        public void FastDicionary_When_Empty_Returns_Enumerable_Of_Zero_Count()
         {
-            IEnumerable<int> dictionary = new FastDictionary<int, int>().EnumerableOfKeys();
+            IEnumerable<KeyValuePair<int, int>> dictionary = new FastDictionary<int, int>();
             That(dictionary.Count(), Is.EqualTo(0));
         }
 
@@ -18,7 +18,7 @@ namespace DevFast.Net.Collection.Tests.Implementations.Concurrent
         [TestCase(63)]
         [TestCase(511)]
         [TestCase(1000)]
-        public void FastDicionary_KeyEnumerator_Reset_Works_Fine(int totalElement)
+        public void FastDicionary_Enumerator_Reset_Works_Fine(int totalElement)
         {
             FastDictionary<int, int> dictionary = new();
             for (int i = 0; i < totalElement; i++)
@@ -26,11 +26,13 @@ namespace DevFast.Net.Collection.Tests.Implementations.Concurrent
                 dictionary[i] = 2;
             }
             That(dictionary, Has.Count.EqualTo(totalElement));
-            using IEnumerator<int> de = dictionary.EnumerableOfKeys().GetEnumerator();
+            using IEnumerator<KeyValuePair<int, int>> de = dictionary.GetEnumerator();
             int count = 0;
             while (de.MoveNext())
             {
-                That(de.Current, Is.LessThan(totalElement));
+                That(de.Current.Key, Is.LessThan(totalElement));
+                That(de.Current.Key, Is.GreaterThanOrEqualTo(0));
+                That(de.Current.Value, Is.EqualTo(2));
                 count++;
             }
             That(count, Is.EqualTo(totalElement));
@@ -38,19 +40,21 @@ namespace DevFast.Net.Collection.Tests.Implementations.Concurrent
             count = 0;
             while (de.MoveNext())
             {
-                That(de.Current, Is.LessThan(totalElement));
+                That(de.Current.Key, Is.LessThan(totalElement));
+                That(de.Current.Key, Is.GreaterThanOrEqualTo(0));
+                That(de.Current.Value, Is.EqualTo(2));
                 count++;
             }
             That(count, Is.EqualTo(totalElement));
         }
 
         [Test]
-        public void FastDicionary_KeyEnumerator_Once_Dispose_Always_Yields_False_On_MoveNext()
+        public void FastDicionary_Enumerator_Once_Dispose_Always_Yields_False_On_MoveNext()
         {
-            using IEnumerator<int> de = new FastDictionary<int, int>
+            using IEnumerator<KeyValuePair<int, int>> de = new FastDictionary<int, int>
             {
                 {1,1}
-            }.EnumerableOfKeys().GetEnumerator();
+            }.GetEnumerator();
             IEnumerator oe = de;
             int count = 0;
             while (oe.MoveNext())
@@ -67,11 +71,6 @@ namespace DevFast.Net.Collection.Tests.Implementations.Concurrent
                 throw new Exception("Enumerator should have returned FALSE");
             }
             That(count, Is.EqualTo(0));
-
-            That(((IEnumerable)new FastDictionary<int, int>
-            {
-                {1,1}
-            }.EnumerableOfKeys()).GetEnumerator(), Is.Not.Null);
         }
     }
 }
