+++ src/DevFast.Net.Collection.Tests/Implementations/Concurrent/Hashed/FastDictionaryTest.cs
@@ -1,8 +1,8 @@
 ﻿using DevFast.Net.Collection.Abstractions;
-using DevFast.Net.Collection.Abstractions.Concurrent.LookUps;
+using DevFast.Net.Collection.Abstractions.Concurrent.Hashed;
 using DevFast.Net.Collection.Implementations.Concurrent.LookUps;
 
-namespace DevFast.Net.Collection.Tests.Implementations.Concurrent
+namespace DevFast.Net.Collection.Tests.Implementations.Concurrent.Hashed
 {
     [TestFixture]
     public class FastDictionaryTest
