+++ src/DevFast.Net.Collection/Abstractions/Concurrent/Hashed/NamespaceDoc.cs
@@ -1,4 +1,4 @@
-﻿namespace DevFast.Net.Collection.Abstractions.Concurrent.LookUps;
+﻿namespace DevFast.Net.Collection.Abstractions.Concurrent.Hashed;
 
 /// <summary>
 /// <see langword="namespace"/> contains necessary <see langword="interface"/>s to implement various hash-based data-structures.
