+++ src/DevFast.Net.Collection/Implementations/Concurrent/Hashed/NamespaceDoc.cs
@@ -1,4 +1,4 @@
-﻿namespace DevFast.Net.Collection.Implementations.Concurrent.LookUps;
+﻿namespace DevFast.Net.Collection.Implementations.Concurrent.Hashed;
 
 /// <summary>
 /// <see langword="namespace"/> contains various hash-based data-structures.
